+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.1.0] - 2020-12-04
 ### Added
 - method `.get_sync_word()`
 - method `.set_sync_word()`
@@ -46,7 +48,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - disabled data whitening
 - automatic calibration
 
-[Unreleased]: https://github.com/fphammerle/python-cc1101/compare/v2.0.0...HEAD
+[Unreleased]: https://github.com/fphammerle/python-cc1101/compare/v2.1.0...HEAD
+[2.1.0]: https://github.com/fphammerle/python-cc1101/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/fphammerle/python-cc1101/compare/v1.2.0...v2.0.0
 [1.2.0]: https://github.com/fphammerle/python-cc1101/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/fphammerle/python-cc1101/compare/v1.0.0...v1.1.0